Chain page refactor, Streamlit telemetry disable
